// src/schemaValidations/category.schema.ts
import { z } from 'zod'

export const CategorySchema = z.object({
    id: z.number(),
    parentCategoryId: z.number().nullable(),
    name: z.string(),
    logo: z.string().nullable(),

    createdById: z.number().nullable(),
    updatedById: z.number().nullable(),
    deletedById: z.number().nullable(),
    deletedAt: z.iso.datetime().nullable(),
    createdAt: z.iso.datetime(),
    updatedAt: z.iso.datetime(),
})